Don't save or report bw results if testing didn't complete before user navigates off page.
If a user navigates away from a page on which bandwidth is being tested
(i.e. before the file download probes are finished), the accuracy of
the results vary widely.

Therefore, don't save or report bw results if testing didn't complete
before user navigates off page.

A side-effect of this change is that bw testing will run on subsequent
pages until is finishes (which only requires that a user remain on a
page for 15s, so valid data should come quickly).